Refactor plansession to MVC
